Added views to list and detail books info and import data from excel

+++ core/models.py
@@ -16,7 +16,7 @@ class Books(models.Model):
     authors = models.CharField(max_length=255, blank=True, null=True)
     publisher = models.CharField(max_length=100, blank=True, null=True)
     published_date = models.DateField(blank=True, null=True)
-    distribution_expense = models.DecimalField(max_digits=2, decimal_places=2, blank=True, null=True)
+    distribution_expense = models.FloatField(blank=True, null=True)
     category = models.ForeignKey(Category, on_delete=models.CASCADE)
 
     def __str__(self):
